requirements.txt
# =============================================================================
# Doom Index Production Requirements (Merged)
# =============================================================================

# --- Core ---
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0
joblib>=1.3.0

# --- Deep Learning / PyTorch (H100 Optimized) ---
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.38.0
datasets>=2.16.0
accelerate>=0.26.0
tokenizers>=0.15.0
safetensors>=0.4.0

# --- DeepSpeed for Multi-Node H100 Training ---
deepspeed>=0.12.0

# --- ONNX / TensorRT Inference Optimization ---
onnx>=1.15.0
onnxruntime-gpu>=1.16.0
onnxruntime-tools>=1.7.0
onnx-graphsurgeon>=0.3.27
tensorrt>=8.6.1
# pycuda>=2022.2.2  # Install separately with CUDA matching toolkit

# --- Data Engineering at Scale ---
pyarrow>=14.0.0
fastparquet>=2023.10.0
polars>=0.20.0
webdataset>=0.2.48
zstandard>=0.22.0
# # simdjson>=5.0.2

# --- Graph Neural Networks ---
torch-geometric>=2.4.0
# torch-scatter>=2.1.2
# torch-sparse>=0.6.18
# torch-cluster
# torch-spline-conv
# pyg-lib
networkx>=3.2.0
neo4j>=5.15.0

# --- NLP ---
nltk>=3.8.0
textblob>=0.17.1
vaderSentiment>=3.3.2
spacy>=3.7.0
ftfy>=6.1.0
regex>=2023.10.0

# --- Computer Vision ---
pillow>=10.1.0

# --- API / Serving ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
httpx>=0.26.0
gunicorn>=21.2.0
celery>=5.3.0

# --- Data Storage ---
pymongo>=4.6.0
redis>=5.0.0
feast>=0.34.0

# --- Feature Store / Monitoring ---
prometheus-client>=0.19.0
psutil>=5.9.6
pynvml>=11.5.0

# --- Streaming ---
kafka-python>=2.0.2
confluent-kafka>=2.3.0
apache-beam>=2.52.0

# --- Privacy / Federated ---
opacus>=1.4.0
flower>=1.6.0

# --- Adversarial / Security ---
textattack>=0.3.10
cleverhans>=4.0.0
cryptography>=41.0.0
bcrypt>=4.1.0

# --- ONNX (already present) ---
# onnx, onnxruntime-gpu already listed above

# --- Visualization ---
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
pyvis>=0.3.1

# --- Dashboard ---
streamlit>=1.29.0

# --- Experiment Tracking / MLOps ---
mlflow>=2.8.0
wandb>=0.16.0
dvc>=3.30.0

# --- Data Validation ---
great-expectations>=0.18.0
pandera>=0.18.0

# --- A/B Testing ---
statsmodels>=0.14.0

# --- Testing ---
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
locust>=2.18.0

# --- Code Quality ---
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
pylint>=3.0.0
bandit>=1.7.5

# --- Utilities ---
pyyaml>=6.0.1
tqdm>=4.66.0
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.0.0
requests>=2.31.0

# --- Profiling / Benchmarking ---
ninja>=1.11.0
